% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A.R
\name{A}
\alias{A}
\title{Compute A}
\usage{
A(hit, fa)
}
\arguments{
\item{hit}{hit rate}

\item{fa}{false alarm rate}
}
\value{
sensitivity A
}
\description{
Compute the non-parameterical measure A for sensitivity
}
\references{
Zhang, J., & Mueller, S. T. (2005). A note on ROC analysis and non-parametric estimate of sensitivity. Psychometrika, 70(1), 203–212. doi:10.1007/s11336-003-1119-8
}
