% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_prime.R
\name{A_prime}
\alias{A_prime}
\title{Compute A'}
\usage{
A_prime(hit, fa)
}
\arguments{
\item{hit}{hit rate}

\item{fa}{false alarm rate}
}
\value{
sensitivity A'
}
\description{
Compute the non-parameterical measure A' for sensitivity
}
\references{
Stanislaw, H., & Todorov, N. (1999). Calculation of signal detection theory measures. Behavior Research Methods, Instruments, & Computers, 31(1), 137–149.
}

