% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b.R
\name{b}
\alias{b}
\title{Compute b}
\usage{
b(hit, fa)
}
\arguments{
\item{hit}{hit rate}

\item{fa}{false alarm rate}
}
\value{
sensitivity b
}
\description{
Compute the non-parameterical measure b for bias
}
\references{
Zhang, J., & Mueller, S. T. (2005). A note on ROC analysis and non-parametric estimate of sensitivity. Psychometrika, 70(1), 203–212. doi:10.1007/s11336-003-1119-8
}
