% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_prime.R
\name{B_prime}
\alias{B_prime}
\title{Compute B''}
\usage{
B_prime(hit, fa)
}
\arguments{
\item{hit}{hit rate}

\item{fa}{false alarm rate}
}
\value{
sensitivity B''
}
\description{
Compute the non-parameterical measure B'' for response bias
}
\references{
Stanislaw, H., & Todorov, N. (1999). Calculation of signal detection theory measures. Behavior Research Methods, Instruments, & Computers, 31(1), 137–149.
}

