% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.R
\name{cc}
\alias{cc}
\title{Compute c}
\usage{
cc(hit, fa)
}
\arguments{
\item{hit}{hit rate}

\item{fa}{false alarm rate}
}
\value{
bias c
}
\description{
Compute the response bias measure c
}
\references{
MacmillanStanislaw, H., & Todorov, N. (1999). Calculation of signal detection theory measures. Behavior Research Methods, Instruments, & Computers, 31(1), 137–149.
}
