% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_prime.R
\name{d_prime}
\alias{d_prime}
\title{Compute d'}
\usage{
d_prime(hit, fa)
}
\arguments{
\item{hit}{hit rate}

\item{fa}{false alarm rate}
}
\value{
sensitivity d'
}
\description{
Compute the parameterical measure d' for sensitivity
}
\references{
Macmillan, N. A., & Creelman, C. D. (2005). Detection theory: A user’s guide. Lawrence Erlbaum Associates.
}
