% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez_pes.R
\name{ezANOVA.pes}
\alias{ezANOVA.pes}
\title{ezANOVA wrapper to calculate}
\usage{
ezANOVA.pes(ezANOVAResult)
}
\arguments{
\item{ezANOVAResult}{ezANOVA result}
}
\value{
ezANOVA with partial eta-squared
}
\description{
ezANOVA wrapper to calculate
}
\details{
ezANOVA muss mit dem Parameter 'detailed=T' ausgeführt werden.
}
\author{
Frank Papenmeier
}
