% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineplot.CI_within.R
\name{lineplot.CI.within}
\alias{lineplot.CI.within}
\title{Lineplot within-subject error bars}
\usage{
lineplot.CI.within(x.factor, response, group = NULL, data = NULL,
  sid = NULL, subset = NULL, conf.level = 0.95,
  within.ci.method = c("baguley", "morey", "cousineau"), ...)
}
\arguments{
\item{x.factor}{a factor (required) whose levels will form the x axis.}

\item{response}{a numeric variable giving the response.}

\item{group}{grouping factor (optional) whose levels will form the traces.}

\item{data}{an optional data frame.}

\item{sid}{subject identifier}

\item{subset}{an optional expression indicating the subset of the rows of 'data' that should be used in the plot.}

\item{...}{further graphical parameters.}

\item{type}{the type of plot: lines, points, or both. Defaults to both.}

\item{x.cont}{logical. Treat x.factor as a continuous variable?}

\item{legend}{logical. Should a legend be included?}

\item{trace.label}{overall legend label.}

\item{leg.lab}{legend labels for trace factors.}

\item{fixed}{logical. Should the legend be in the order of the levels of 'trace.factor' or in the order of the traces at their right-hand ends?}

\item{x.leg, }{y.leg optional values to over ride the default legend placement.}

\item{cex.leg}{character expansion value for legend labels.}

\item{ncol}{number of columns to use for legend.}

\item{pch}{a vector of plotting symbols or characters.}

\item{fun}{the function to compute the summary statistic. Should return a single real value. Defaults to mean with NA values removed.}

\item{ci.fun}{the functions to compute the CI. Should return a vector of length 2 defining the lower and upper limits of the CI's. Defaults to the mean +/- 1 standard error, with NA values removed.}

\item{err.width}{set width of whiskers for error bars.}

\item{err.col}{color for error bars. Defaults to col.}

\item{err.lty}{line type for error bars.}

\item{xlim, }{ylim range for x and y axes.}

\item{cex}{overall plot character expansion value.}

\item{lwd}{deterimines line width.}

\item{col}{default color(s) for plot.}

\item{cex.axis}{character expansion value for axis labels.}

\item{xaxt}{should x-axis be drawn?}
}
\description{
Lineplot within-subject error bars
}
\author{
Frank Papenmeier
}

