% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.loss_x_y.R
\name{lm.loss_x_y}
\alias{lm.loss_x_y}
\title{Estimate the parameters of a linear regression (intercept, slope, error standard deviation) 
with maximum likelihood estimations}
\usage{
lm.loss_x_y(par)
}
\arguments{
\item{par}{parameter triple (slope, intercept, error standard deviation)}
}
\value{
deviance
}
\description{
Maximum likelihood estimation: Linear regression with one predictor
}
\references{
https://rpubs.com/YaRrr/MLTutorial
}
