% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pes.R
\name{pes}
\alias{pes}
\title{Calculate partial eta squared of a given F statistic}
\usage{
pes(DFeffect, DFerror, Feffect)
}
\arguments{
\item{DFeffect}{Degrees of freedom of the effect}

\item{DFerror}{Degrees of freedom of the error}

\item{Feffect}{F value of the effect}
}
\description{
Calculate partial eta squared of a given F statistic
}
\author{
Markus Huff
}

