% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pes_to_f.R
\name{pes_to_f}
\alias{pes_to_f}
\title{Converts partial eta squared to Cohen's f}
\usage{
pes_to_f(pes)
}
\arguments{
\item{pes}{partial eta sqaured}
}
\description{
Converts partial eta squared to Cohen's f
}
\author{
Markus Huff
}
