% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sciplot-within-CI.R
\name{sciplot.CI.within.common}
\alias{sciplot.CI.within.common}
\title{Sciplot within-subject error bars}
\usage{
sciplot.CI.within.common(
  x.factor,
  response,
  group,
  data,
  sid,
  subset,
  conf.level,
  within.ci.method
)
}
\description{
Implemented within CI Methods: 
Cousineau-Morey intervals with adaption by Baguley (2012), Morey (2008), and Cousineau(2005)")
}
\author{
Frank Papenmeier
}
